+++ src/Event/ContextCreatedEvent.php
@@ -1,0 +1,12 @@
+<?php
+
+namespace Castor\Event;
+
+use Castor\Context;
+
+class ContextCreatedEvent
+{
+    public function __construct(public readonly string $contextName, public Context $context)
+    {
+    }
+}