+++ CHANGELOG.md
@@ -6,6 +6,10 @@
 
 * Fix completion when update is available
 
+### Features
+
+* Add `Castor\Event\ContextCreatedEvent` to allow updating the context after it is created
+
 ## 0.17.1 (2024-05-31)
 
 ### Fixes